Put constants and globals right after imports
It is often recommended to define module constants at the top of
the module or file, directly after the import statements.
Constants are, well, constants and used as basis, configuration,
and components of the code in functions and classes, so it makes
sense to define them first.  If a module uses globals, they may
follow right after the constants, as they are used in the whole
module, too, but may depend on the constants.

Therefore let’s move our definitions of constants and globals
to the top of the file and insert them right after the imports
and before the class definitions.  This shows their importance
and role, and it also resolves the somewhat misleading issue
that until now our function definitions were interrupted by
the definitions of constants and globals.

NB: This commit just moves the complete code block with constants
and globals to the top of the file, inserting them right after
the import statements.  There is only a single change to the code:
in the definition of `PREFIX_REGEX`, the use of `wikipattern()`
is replaced by a simple call to `str.replace()` because
(i) obviously we can’t use our functions here yet, and
(ii) we need indeed just a single replacement operation.
The commit does not change anything else, and the functionality
of the bot program is not changed at all.

Please note that common `diff` tools, including Gihub’s comparison
view of the commit, display this incorrectly, suggesting changes
to the code of several functions, too, but there are no changes.