Implement ‘-test’ task for delist nominations
Implemented the `-test` task, i.e. the recounting and comparison
of the votes, also for delisting nominations.  This feature isn’t
the most important one, of course, but it was an irritating gap
in the program and is indeed quite useful for testing.
And the code has grown only by 8 lines ... that’s fair, isn’t it?

Of course the new code doesn’t work with delist-and-replace
nominations.  Normally it will ignore them, saying a nomination
‘contains alternatives’; but if the nominator uses some unusual
formatting, e.g. a `<gallery>` element, to group the old and
the new image, the program will just report a failed recounting.
But that’s not bad, considering that the bot does not support
new delist-and-replace nominations, too ;–).  I will soon add
that missing feature, too; it’s already half-finished, and for
the testing of the implementation it’s a necessary prerequisite
that the bot program is able to recount delisting nominations.

Polished the related methods `isDone()` and `isPassed()`, and
improved their docstrings.  Improved the names of some local
variables etc. to make the code easier to understand.
Changed the formatting of the output a bit to make it clearer
and to make it more similar to the results of the `-info` task.
Improved the method `existingResult()` by filtering the text
for commented, striken, etc. passages before searching for the
old result, just as we do in `countVotes()` etc.

One statement in the new code uses very clumsy programming style
(checking the class with `isinstance()` instead of relying
on inheritance and polymorphism).  I will replace this soon
by a simple and clean implementation, but in order to do that
without unnecessary expenditure I must first refactor some parts
of the existing code.