Refactor, improve addAssessments()/removeAssessments()
The two methods `FPCandidate.addAssessments()` and
`DelistCandidate.removeAssessments()` modify the `{{Assessments}}`
template in image description pages; both of them add or change
the `featured` parameter and add/update the `com-nom` parameter.
Therefore it makes sense to move the shared code to a new function,
`update_assessments_template()`, which is called by both methods.
This new function also improves several aspects of the previous
implementations:

1. It applies one and the same precompiled regex constant in order
to find existing `{{Assessments}}` templates.  This regex uses `\s`
(instead of simple spaces) and `re.DOTALL`, therefore it tolerates
also templates which have been broken into several lines.
This fixes a real problem, of course, because sometimes people break
the wikitext of this template into several lines and add fancy
whitespace formatting (so-called pretty printing).  In the past
FPCBot used to add a second (!) `{{Assessments}}` template to pages
with such multi-line `{{Assessments}}` templates, that was a bug.

2. Instead of first removing any existing `featured` and `com-nom`
resp. `subpage` parameters and appending new parameters at the end
of the `{{Assessments}}` template, we extract the parameters,
try to update the values of the `featured` and `com-nom`/`subpage`
parameters, and append each of them only if it is missing.
This allows us to preserve special whitespace in the template
perfectly (see above).  BTW we also update the obsolete `subpage`
parameter name to the modern `com-nom` name.

3. The new code handles simple invalid values for the `featured`
parameter, e.g. `true`, gracefully and replaces them by the valid
new value.  This is not very important, of course (I think I have
seen a faulty `featured=yes` once), but it won’t do any harm and
is certainly better than the traditional behaviour of the bot
program which would have just added a second `featured` parameter
if the value of the original one was not a sequence of digits.

4. If there is already an `{{Assessments}}` template, but without
a `com-nom`/`subpage` parameter, the new code takes care to insert
the new `com-nom` parameter right after the `featured` parameter
because the two parameters belong together.

5. Finally, we replace only the parameters, not the complete
`{{Assessments}}` template, again in order to retain the formatting
of the template wikitext.

Changes specific to `FPCandidate.addAssessments()`:

1. Droped the superfluous local variable `current_page`.

2. If the image to which we want to add an `{{Assessments}}` template
is missing (because the image has been deleted or renamed without
a redirect), we print an error to the CLI/log file and continue with
the next image (if we are processing a set nomination).
If the image was renamed with a redirect, we try to resolve that
redirect -- it would be useless and even irritating to add an
`{{Assessments}}` template to the redirect page where nobody will
ever notice it.  (FPCBot used to do this from time to time ;–).)
If resolving the redirect doesn’t work, we print an error message.
If the image is locked so that we cannot commit the new text,
we print an error message, too.  (We do not call `ask_for_help()`
here because we will probably soon unite all image file tests in
a common method.)

3. Rewrote the code which searches for the best place to add a new
`{{Assessments}}` template.  The function `findEndOfTemplate()`
accepts a regex parameter, so we neither need to search first
for the presence of a template with a normal `re.search()`, nor
do we need to search for all the different templates one by one.
We just need two searches, one for the location templates (giving
them priority because they usually appear after the info templates
and are displayed as an unity with them), one for the common
information templates; each search combines the template names
to a single regex with `|` and `(?:...)?` alternatives.
This simplifies the whole search a lot, and allows us to support
easily misc. variants of the location and information templates.
Therefore the commit also adds support for the outdated, but still
used template `{{Location dec}}` and for the common info templates
`{{Artwork}}`, `{{Photograph}}`, and `{{Art photo}}`.

4. But before searching for these templates, we first check
if we can just put the `{{Assessments}}` template at the very
end of the file description stuff by putting it right before
the header of the license section.  This works very well with
most newer files on Wikimedia Commons because most files use
that arrangement of the parts of the file description page;
it is created automatically by the Upload Wizard and was already
used in the default text of the older Upload form.

5. The old code always added a NL after a new `{{Assessments}}`
template; but this often resulted in two (!) empty lines after it
because the template is typically placed at the end of the file
description section which is already separated by one empty line
from the following license information section.  Two empty lines
in the wikitext look a bit odd; therefore the new code takes care
that there is no empty line before and exactly one empty line
after a new `{{Assessments}}` template.

6. Improved the structural comments of the method in order
to help future maintainers to understand the code more easily.

Changes specific to `DelistCandidate.removeAssessments()`:

1. United the three calls of `re.sub(...)` which are used to remove
‘Featured pictures of/from/by ...’ categories to a single call
with a single precompiled regex constant.  This makes the code
both faster and cleaner.

2. If updating the `{{Assessments}}` template does not change it
(this can happen if the parking procedure has previously been
interrupted), we do not need to commit the text of the description
page, but just print a ‘Skipping ...’ note to the CLI/log file.

3. Now we use a more specific error message if the file description
page is locked and cannot be updated.