Improve and document findEndOfTemplate()
Improved the docstring of the function `findEndOfTemplate()`,
documenting the function parameters and explaining that it can also
handle several alternative template names at once if you pass them
as a regex fragment.

Added `\s*[|{}]` at the end of the regex to fix a nasty little bug:
until now the search for `[Ii]nformation` stopped also at instances
of `{{Information field|...}`, but that was wrong, because that
template is used as a component for `{{Information}}` templates,
so the returned end position was right within the `{{Information}}`
template and not at its end.  To avoid this and any similar problems
we must make sure that the template name is treated as a complete
name, and this is achieved by the `\s*[|{}]`.

Escaped the opening braces in the regex properly to avoid potential
conflicts with other/future regex engines and versions.

Improved the initial search by starting it not after the opening
`{{` of the template, but right at `match.end(1)`, i.e. at the
first character after the name of the template and after any
optionally following whitespace.