Improve output formatting for ‘-info’, ‘-test’ tasks
1. The output of the `-info` task uses the labels ‘S’, ‘O’, ‘N’
for the columns with vote counts.  This is a bit misleading:
while ‘S’ and ‘O’ perfectly fit normal FP nominations which use
‘support’ and ‘oppose’ votes, they do not fit delisting nominations
which use ‘delist’ and ‘keep’ votes.  Therefore let’s change
these labels to the more general abbreviations ‘P’, ‘C’, ‘N’,
i.e. ‘pro’, ‘contra’, and ‘neutral’, just as we already use them
in the output of the `-test` task.

2. The two column labels ‘D’ and ‘De’ are easy to confuse; let’s
change ‘D’ to ‘Do’, this is a tiny mnemonic aid that the column
is for the ‘days old’ value and also helps to distinguish it more
clearly from the preceding three vote count columns which all use
one-letter labels.

3. The tables printed by the `-info` and `-test` tasks both contain
a column with boolean values, but in one case we print them as
‘True’ and ‘False’, in the other case as ‘1’ and ‘0’.  This is
a bit confusing, of course.  Using ‘1’ and ‘0’ saves as much space
as possible, but is also a bit unclear because all other columns
really integers, so ‘1’ and ‘0’ are easily confused with them.
Using ‘True’ and ‘False’ is clear, but needs unnecessarily much
space in the column when we consider that this column does not
represent the result, but just one value among many other ones.
I thought about abbreviating the column value with ‘T’ and ‘F’,
as usual in truth tables, but unfortunately these letters look
quite similar in the terminal, so it’s hard to spot an outlier.
And the Wiki world has an interesting preference for using
‘yes’ and ‘no’ for boolean parameters, not ‘true’ and ‘false’.

So let’s use the abbreviations ‘Y’ and ‘N’ for boolean values
in the tables printed by the `-info` and `-test` tasks.
This is short, keeps the table nicely aligned, doesn’t interrupt
the tables, but still allows to spot outliers etc. easily because
the letters ‘Y’ and ‘N’ look so different.

4. The method `Candidate.printAllInfo()` clearly shows the limits
of the classic `%` operator for string interpolation: with no less
than 10 variables, it is really hard to see which value will end
where in the final string.  IMHO using f-strings makes the code
much clearer and easier to understand, as it makes obvious
which variable is interpolated after which label and with which
format specification.  Likewise the simple error messages in
`compareResultToCount()` profit from using f-strings because they
put the variable right to the place to which they belong and not
at exactly the other end of the line, as the use of `%` did.