Fix bug in auxiliary image search
The method `Candidate.fileName()` infers the name of the nominated
image file from the name of the nomination subpage.  If there is
no file with that name, it tries to use the name of the first image
which is linked from the nomination subpage instead.  Until now
the search for that image was buggy because it did neither ignore
commented, stricken out, etc. text (which may very well contain
image links, but certainly not the nominated image) nor did it
check the details of the image link in order to sort out mere icons
and tiny thumbnails, as we do in `Candidate.imageCount()`.

Moved the image checking code from `Candidate.imageCount()` to a new
simple utility function `is_just_thumbnail()` which is now called
both from `Candidate.imageCount()` and from `Candidate.fileName()`.
Improved the code of the latter method by filtering the wikitext
of the nomination subpage first with our `filter_content()` function
in order to ignore comments, code sample, stricken text, etc.