Decode %-encoded chars in gallery links
From time to time a gallery link contains a `%xx` escape, normally
for ‘#’ (`%23`) or ‘:’ (`%3A`).  Such gallery links don’t work,
of course, because we treat the gallery links as plain text and
fail to identify the `%3A` as section anchor separator, etc.
Taught the bot to decode `%xx` escapes when cleaning gallery links.