Cache filtered version of nomination subpage text
Let’s cache the filtered version of the nomination subpage text,
i.e. the redacted version without comments, <s>...</s> passages,
code examples, etc.  We need this version quite often, and it’s
useless to filter one and the same text again and again.

We cache only the filtered version of the nomination text because
in almost all cases we need the filtered version, and in the rare
cases when we need the full text of the subpage in order to edit
it we must always get the most recent version, of course.

The commit also makes sure that we use the filtered text in all
places which benefit from removing comments etc.; this improves
the accuracy of some methods further, see e.g. `setFiles()` or
`findGalleryOfFile()`.  At the same time speed is improved, too,
because almost everywhere we now just re-use the filtered text
which is cached and shorter, so even faster to search.

The commit also gets rid of some related disposable (single-use)
local variables, improves the names of some variables, etc.

NB: This commit does not change the error handling related to
getting or saving the contents of the nomination subpage.
Everything is as before, with the exception that now we always
access the contents of the nomination subpage with Pywikibot’s
`get_redirect` flag set to `False`; that’s intentional because
we *never* want to get a mere redirect, but always need the real
content.  This should not make any practical difference because
`findCandidates()` already checks all nomination subpages for
redirects and resolves them before we process them, but it is
good to make sure that we are never taken for a ride by a mere
redirect page.  I hope to tackle further improvements for the
error handling soon.