Access Pywikibot page.Revision data items by keys
The data items of Pywikibot’s class `page.Revision` can be accessed
either by their keys or as attributes with the same names.
Until now whe have used the access via attributes, but it’s
difficult to get proper type hinting and checking with this approach
because Pywikibot doesn’t define the types of all these attributes
or virtual attributes.  Therefore mypy rightly reports the following
issues:
    "Revision" has no attribute "timestamp" [attr-defined]
    "Revision" has no attribute "user" [attr-defined]
The easiest and clearest way to solve this is to access the data
items by their keys.