Introduce type hinting
Introduce type hinting for all module constants, module variables
(globals), instance variables (attributes), function and method
signatures, and annotate a few difficult local variables.

This is enough to allow mypy (or similar typecheckers) to check
the types of all items in our module, and it is also a great help
for human code readers.  I have missed type hints since the first
day I took a look at the code of this project, so please allow me
to amend it, finally, with modern type hinting.  This will help us
a lot to prevent the introduction of new errors and glitches,
because now mypy & Co. can analyse the code thoroughly,
and it will also help future maintainers to understand and edit
this project much more easily.