Use class constants for constant class-specific values
The `Candidate` class uses a lot of instance variables.  Some of them
are very much like constants: they are defined by the initializer
with the appropriate value for a certain subclass (FP candidate vs.
delisting candidate), after that the value never changes.  This means
the value always stays one and the same during the lifespan of any
class instance; it does not depend at all on the individual instance,
only on the concrete subclass.

This suggests to move the definition of these values from the
instance level to the class level.  We can clarify the structure
and usage of the `Candidate` class if we simply define them as
class attributes instead of instance variables.  This makes obvious
that the meaning and values of these attributes are related only
to the concrete subclass, not on the individual instance.  It also
allows us to simplify the initializers (`__init__()`) both of the
base class and of the subclasses drastically.  The rest of the code
is just updated to use the new names; there are no disadvantages
for speed or memory usage, on the contrary: both should even
benefit a little bit from this change.

Further simplifications which are made possible by this change
will follow in the next commits.

NB: I describe the new class attributes as class constants and use
the conventional UPPER_CASE_WITH_UNDERSCORES naming style for them
because the values of these attributes never change at runtime;
IMHO this is the specific feature of a constant, and using
the conventional naming style allows linters to verify this.
If, however, there arises a consensus that we should not consider
these class attributes as class constants (because their values
are adapted by a subclass at definition time), it is very easy
to change this: one only needs to change the naming style
to the conventional lower_case_with_underscores and to update
the comments, then all these values will be considered as class
variables, the rest of the code doesn’t need any changes.