Drop __init__() methods from subclasses
Thanks to the last changes to the `Candidate` class the `__init__()`
methods (initializers) of the subclasses just call the superclass
initializer.  This means we can drop them from the subclasses and
manage with the superclass initializer; Python automatically
calls it if there is no subclass initializer.

We will need the subclass initializers again if and only if we ever
add any instance variables to the subclasses which are specific to
these classes (or other subclass-specific stuff which needs to be
set up by the initializer).