Sort definitions of instance variables
Sort the definitions of the instance variables of the `Candidate`
class both in the `__init__()` method and in the declaration of
types for these instance variables.  The new order follows
as far as possible the meaning of these values, groups related
variables and therefore helps to understand the class easily.