Add/improve comments on class consts, instance vars
I hope to improve the docstrings soon in order to document functions,
classes, and class methods even better.  But internal attributes
like internal class constants and internal instance variables are
usually not documented in docstrings because they don’t belong to
the public API.  Therefore it makes sense to add a short comment
to each of them in order to help future contributors to understand
immediately what each of these values means and what distinguishes
some similar values from each other.

Exchanged the order of the class constants for counted (unreviewed)
and reviewed results because the bot program first adds the counted
(unreviewed) results and only after that they are reviewed by a human
editor and changed to reviewed results.