Fix handling of badly named nomination subpages
If the name of a simple FP nomination subpage does not follow our
usual format, i.e. is not of the form ‘Commons:Featured picture
candidates/<filename>’ with `filename` including the ‘File:’ or
‘Image:’ namespace prefix, FPCBot cannot properly derive the name
of the nominated image and also does not stop on that error.
Hence our poor bot fails miserly because it tries to treat the
nomination subpage as the nominated image.  This has become obvious
in a recent nomination [1].

We certainly can do better here.  The new code makes sure that any
failure to derive the image name from the name of the nomination
subpage is recognized.  As a workaround it then tries to derive
the name of the nominated image by searching for the first preview
image in the nomination text (just as we already do when the name
of the image was derived properly, but no image with that name
exists on Commons).  This is sufficient to fix the bot’s confusion.

At this opportunity I have also rewritten the error handling in
`Candidate.filename()`.  We now report right from that method
if we cannot find the nominated image resp. if it does not exist.
This allows us to drop a similar test in `Candidate.park()` and
may allow us to avoid some other tests elsewhere when we find
the time to restructure the related code.  Already right now
it improves the error messages (we now have a special message
if the name of the nomination subpage is irregular) and makes sure
that we skip the rest of the parking procedure for the current
nomination if we cannot identify the nominated image, but can
continue properly with the next candidate from the list.
This means the bot does not stop anymore completely if it ever
encounters such an error.

This commit requires a little change to the handling and semantics
of the instance variable `Candidate._filename’: now we define
that variable as empty if we cannot identify the image properly.
This means we can distinguish whether the variable has never been
set (then its value is still `None`) or if it was set, but only
in order to mark it as ‘not found’ (then its values is just
an empty string).  This helps to avoid useless repeated attempts
to acquire the filename of the nominated image.

[1] https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/Varanus_komodoensis,_Komodo_Island,_Indonesia,_20250822_1332_2808.jpg