Handle Candidate._alternative like Candidate._filename
Handle the instance variable `Candidate._alternative` just like
we now handle `Candidate._filename`: define the variable as empty
if we cannot identify the selected alternative image properly.
This means we can distinguish whether the variable has never been
set (then its value is still `None`) or if it was set, but only
in order to mark it as invalid/not found (then its value is just
an empty string).  Right now this does not make any practical
difference for `_alternative`, but it helps to avoid confusion
by making sure that the related instance variables `_filename`
and `_alternative` are also handled similarly.

Move the declaration of `_alternative` right after the declaration
of `_filename` to make their relation and similarity obvious.