Improve redirect handling in Candidate.filename()
Add proper exception handling to the code in `Candidate.filename()`
which resolves redirects to renamed candidate images.  If the bot
ever encounters a broken or circular redirect (unlikely, but not
impossible in the wonderful world of Commons), it will now report
this properly on the FPC talk page and stop to process the current
nomination, continuing with the next one from the candidate list.

Add two warnings about the need to derive the filename from the 1st
image in the nomination; if something goes wrong that warning will
appear in the log file right before any errors and help us to
understand what is going one.