Allow comment between section heading and <gallery>
To avoid the misplacing of new FPs into wrong gallery sections
(which was quite common with the old bot code), we require that
the target subheading is followed *immediately* by the associated
`<gallery>...</gallery>` element.  But we tolerate a one-line
‘See also ...’ link etc. between the two, making sure that this
line does not start with a `=` or `<` because any additional
subheading or `<gallery>` element would break the required strong
connection between target subheading and `<gallery>` element.

A few gallery pages (e.g. [1]) use HTML comments between some
subheadings and their `<gallery>` elements [1].  Such comments
make sense and can be very useful, of course, so the bot must
tolerate them.  This commit updates the comment on the search,
improves the formatting of the regex code and amends it by adding
a new exception for HTML comments.  Unlike other stuff we allow
such comments also to span several lines.

[1] https://commons.wikimedia.org/wiki/Commons:Featured_pictures/People/Work