Fix duplicate notification due to spelling difference
The Mediawiki software handles the first character of page names
and user names (after the namespace prefix) case-insensitively,
but the remaining characters case-sensitively.  This allows for
some variation in the spelling of page links and also user names
which can have confusing consequences.  If the first letter of
the name in a `[[User:...]]` link is changed, Mediawiki still
recognizes and handles the link correctly, but our program fails
to understand that creator and nominator of an image are identical
and adds two FP notification templates to the talk page of the
same user.   This has happened e.g. with the nominations [1], [2].
That’s a bit confusing and maybe even impolite, so let’s try
to avoid it.  There are two simple things we can do two things
in order to improve the comparison of usernames:

1. Replace `_` by ` ` when extracting the name of the original
creator of a nominated image from the Wikitext.

2. Add and use a special comparison function which compares two
usernames in a special way, handling their first letters case-
insensitively, just as the Mediawiki software does.

[1] https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/File:Bingen_Toll_Tower_-_3.jpg

[2] https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/File:M42_(infrared,_S-II,_O-III).jpg