Always mention creator in chronological entries
In the entries of the chronological archives, we always mention
uploader and nominator separately, even if they are one and
the same user, in order to keep the traditional format of the
chronological archives as far as possible and to simplify
statistical evaluations.  It makes sense to use the same approach
for the name of the creator, too: if we like repetition and want
to be completely explicit, this should apply to the creator name,
too ;–).  Therefore let’s mention the creator whenever we can
identify them, even if the username is the same as the name of
the uploader.

If we would keep the test if creator and uploader are different,
we would have to rewrite it to use our new function
`is_same_user()`, because a simple `!=` test is not sufficient
(see previous commit).  Changing our policy to always include
the name of the creator (if available) avoids this, too.