Handle errors, resolve redirects right in set_files()
Recently the handling and resolving of redirects in the method
`Candidate.filename()` has been improved, this commit does more or
less the same for `Candidate.set_files()`.  Now we try to detect
missing files, redirects, circular or broken redirects right in
that method.  Together with the aforementioned changes to
`Candidate.filename()` this means that we apply all these tests
right in the methods which retrieve the name(s) of the nominated
image(s), so we can remove any tests of this kind from other
methods.

Furthermore, this commit improves error handling and error reports.
It updates the semantics of the instance variable `_set_files`:
similar to the new semantics of `_filename` and `_alternative`,
`None` now means that the variable has not been set yet, while
an empty value means that we have tried to retrieve the filename(s)
but that this was not possible due to a serious errors.  This allows
to distinguish clearly between the initial value (`None`) and the
value which is used in cases of errors.