Drop nominator ping feature
Pinging the nominator from a request for help seems not to work;
either the users do not get the ping at all (because of the bot
flag?), or they happily ignore it.  Well, that’s a sad story,
but at least it allows us to drop that feature again, reducing
the increasing complexity of the bot program a tiny little bit ;–).