Improve gallery link error messages
Seems that the messages which the bot posts on the FPC talk page
about invalid gallery links are still hard to understand and not
sufficiently helpful for the nominators and other editors.
This commit refactors `FPCandidate.add_to_gallery_page()` and
some related code in order to gain more solid information about
the exact problem with the gallery link, and uses it to generate
more concrete warnings.  The wording may still need improvements,
but all in all the new messages should be much clearer and more
helpful.

Of course we have now reached the point where more than half of
the code of that method is required only in order to provide
nice error messages and warnings.  That seems a bit unbalanced.
But in the end FPCBot does not need to be an example of especially
elegant and concise design, it is just a bot program which should
assist the human editors with the management of the FP candidates,
and save them as much work as possible.  Therefore it seems OK
to dedicate an unusual amount of code just to the generation
of clear error descriptions and helpful warnings.