Warn if users specify ‘Unsorted’ as section anchor
Sometimes users actually use `#Unsorted` for the section anchor
in a FP gallery link [1].  That is not exactly helpful because
this section is used only for images which need to be sorted
into a more specific section.  In all other cases in which the bot
adds one or more new FPs to the ‘Unsorted’ section it posts
a request for help to the FPC talk page.  In consequence the bot
should also ask for help if a user specifies `#Unsorted` as
section anchor.  So the new PF(s) are added to the ‘Unsorted’
section, but the editors are notified and can move the images
to a more appropriate place in the FP gallery pages.

[1] E.g. in this nomination:
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/File:Arenal_volcano_(70785p).jpg