Fix success value in -info output for noms with alts
As of a short while ago the output of the `-info` task contains
a column with the current success tendency for each nomination --
‘Y’ or ‘N’ for ‘yes’ or ‘no’.  For nominations with alternatives
this value was misleading and often wrong because our bot program
is not (yet) able to count the votes in such complex nominations
properly.  Therefore let’s better print just ‘?’ in the column
with the current success tendency for such nominations.