Move search for gallery section to new method
The method `FPCandidate.add_to_gallery_page()` has become very long
and somewhat unclear.  We can clarify the structure of the method
by moving the search for the place to insert the new FP(s) to a new,
separate method.  The name of that method can be prefixed with `_`
because it is just an internal subroutine which must not be called
from outside of the class.  The flow of control in the new method
is intentionally super simple and boring to allow future contributors
to understand and to modify the code easily.