Use slice instead of range for insertion place
In the refactored code of `FPCandidate.add_to_gallery_page()` and
in its new subroutine I used a Python `range` object to describe
the index values of the characters of the gallery page wikitext
which should be replaced by the new gallery entries.  It seems
a bit more fitting to use a Python `slice` object instead because
we describe the indices of a certain span of characters.
Likewise, it makes sense to use more descriptive method, variable
names and to improve the docstring of the new method.