Insert FPs at the top of ‘Unsorted’ section
Since August 2025 (GitHub commit 920fc63), we insert all new FPs
no longer at the end, but at the top of a gallery section.
It seems just consequential to do the same when we add FPs
to the ‘Unsorted’ section, i.e. to insert them at the top of the
`<gallery>...</gallery>` element for that special section.
This commit implements that idea, and also adds the removal
of any unnecessary/unwanted whitespace before and after the new
entry/entries, just as we already do when inserting any new FP(s)
into normal gallery sections.  The commit also moves the code
for the handling of the ‘Unsorted’ section to a new subroutine,
making `Candidate.add_to_gallery_page()` shorter and clearer.