Update self-links in nomination subpage after renaming
‘Self link’ is the usual MediaWiki term for any wikilink which points
to the page on which the link is placed.  All typical FP nominations
contain several self-links.  If we rename a nomination subpage,
it therefore makes sense to update these self-links immediately.
(We could do without that because on renaming we create a redirect
which keeps these indirect self-links functional, but it’s certainly
much cleaner to update all self-links right away.)

In line with that it also makes sense to update any links to the
nominated image in the nomination subpage, because these links
will often contain the same wrong spaces etc. as the self-links.

Finally, there is an especially tricky link in every normal
FP nomination: the span
`<span class="wpImageAnnotatorPageName" style="display:none;">`
contains yet another self-link to the nomination subpage, but
this time *without* the `Commons:` namespace prefix of the subpage.
(Don’t ask me why this is the way it is or why this was designed
in that odd way.  The ways of the Wiki world are inscrutable.)
In order to catch and to update this crazy link, too, we use the
alternative `(\[\[:?|/)` at the beginning of the regex for updating
the file links: the `\[\[:?` is for links to the nomination image,
the `/` is for that very special `wpImageAnnotatorPageName` link.

NB: Delisting nominations normally don’t contain any self-links,
therefore we do not report an error if no self-links are found
and replaced.