Report completely invalid nomination names
The most problematic nomination subpage names are those which
(i) do not even start with the obligatory part ‘Commons:Featured
picture candidates/...’, or (ii) do not even clearly indicate
whether the nomination is a simple single-file nomination or
a set nomination because their names contain neither the namespace
prefix `File:` (or `Image:`) nor the signal word `/Set/` [1].
Therefore even the type of such nominations is unclear.
Deriving it from the contents of the subpage should be possible,
but (1.) would be cumbersome, (2.) may yield wrong results if the
subpage uses unusual formatting, and (3.) would require us to read
the nomination subpage even before constructing the `FPCandidate`
object for it.  So we better draw a line here.

We have only recently improved the way the bot handles such
nominations in order to avoid severe damages to the FP gallery
and archival pages (GitHub commit 9545ebc, ‘Fix handling of badly
named nomination subpages’).  But we better add yet another
level of ‘defense’ against such malformed nominations and
reject them immediately, reporting them on the FPC talk page.
This allows the FPC regulars to fix the nominations right away
before the bot closes and parks them.

[1] The most recent example for this is
https://commons.wikimedia.org/wiki/Commons:Featured_picture_candidates/Varanus_komodoensis,_Komodo_Island,_Indonesia,_20250822_1332_2808.jpg