Test ‘-match’ argument against full subpage name
We used to compare the value of the `-match` argument only with
the very last part of the name of the nomination subpage, skipping
not only the ‘Commons:Featured picture candidates/’, but also
the ‘File:’ or ‘Image:’ namespace prefix and anything between them,
i.e. any superfluous spaces and the signal word ‘/removal/’.

But there was a little bug: the regex we used to remove the leading
part of the name worked only if the subpage name contained the
‘File:’ or ‘Image:’ namespace prefix; it did not work at all with
set nominations, so for set nominations we always compared
the `-match` argument with the full name of the nomination subpage.
That’s not a severe problem, of course, therefore that bug was never
noticed, but it is nevertheless a nasty inconsistency.

Furthermore you may sometimes want to include the ‘/removal/’ or
‘/Set/’ signal word in your `-match` argument, to find only
nominations of that specific type, or even the ‘File:’ namespace,
e.g. to catch specifically nominations about files whose name starts
with a certain letter like ‘A’ -- in that case `-match File:A`
will do the trick.  This is not possible as long as we remove
these parts of the subpage name before testing it.

We can solve both issues at once if we remove only the common part
of *every* nomination subpage name, i.e. the ‘Commons:Featured
picture candidates/’, before comparing it with the `-match` value.